+++ integrations/tensorflow/e2e/BUILD
@@ -56,8 +56,8 @@ package(
 SPECIAL_CASES = [
     "explicit_backend_test.py",
     "linspace_test.py",
-    "simple_arithmetic_trace_test.py",
-    "strings_trace_test.py",
+    "simple_arithmetic_test.py",
+    "strings_test.py",
 ]
 
 # keep sorted
@@ -210,13 +210,13 @@ iree_py_test(
 # prototyping a change.
 
 iree_py_test(
-    name = "simple_arithmetic_trace_test",
-    srcs = ["simple_arithmetic_trace_test.py"],
+    name = "simple_arithmetic_test",
+    srcs = ["simple_arithmetic_test.py"],
     args = [
         "--reference_backend=tf",
         "--target_backends=tf,iree_vmla,iree_llvmjit",
     ],
-    main = "simple_arithmetic_trace_test.py",
+    main = "simple_arithmetic_test.py",
     python_version = "PY3",
     tags = [
         "driver=llvm",
@@ -228,13 +228,13 @@ iree_py_test(
 )
 
 iree_py_test(
-    name = "strings_trace_test_tf",
-    srcs = ["strings_trace_test.py"],
+    name = "strings_test_tf",
+    srcs = ["strings_test.py"],
     args = [
         "--reference_backend=tf",
         "--target_backends=tf",
     ],
-    main = "strings_trace_test.py",
+    main = "strings_test.py",
     python_version = "PY3",
     deps = INTREE_TENSORFLOW_PY_DEPS + NUMPY_DEPS + [
         "//integrations/tensorflow/bindings/python/pyiree/tf/support",
